Move service api to each pages api folder
