--- client/app/utils/api/index.ts
@@ -1,2 +1,0 @@
-export * from './category'
-export * from './todolist'
