+++ client/app/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './customFetch'
-export * from './api'
 export * from './mouseEvent'
 export * from './time'
